+++ app/(main)/c/[id]/page.tsx
@@ -1,14 +1,13 @@
 import {
-  CourseInfo,
-  CreateTest,
-  ShortLearderboard,
-  TestBox,
+    CourseInfo,
+    CreateTest,
+    ShortLearderboard,
+    TestBox,
 } from '@/components/Course';
 import { CreateTestPopUp, EditCourse, FullScreenPopUp } from '@/components/Popups';
 import { CourseNavFiller, MainNavFiller } from '@/components/Utils';
-import axios from 'axios';
 import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
+import { fetchCourseInfo } from './action';
 
 interface Course {
   id: number;
@@ -23,40 +22,7 @@ interface Course {
   leaderboard: { name: string; points: number }[];
 }
 
-export async function fetchCourseInfo(id: number) {
-  const cookieStore = cookies();
-  if (!cookieStore.get('access_token')) {
-    redirect('/auth/signin');
-    return {
-      props: {
-        data: null,
-      },
-    };
-  }
 
-  try {
-    const response = await axios.get(`http://localhost:5000/course/${id}`, {
-      headers: {
-        Authorization: `Bearer ${cookieStore.get('access_token')?.value}`,
-      },
-    });
-
-    const data = response.data;
-    return {
-      props: {
-        data,
-      },
-    };
-  } catch (err: any) {
-    //todo: error handling
-    console.error('Error fetching course data:', err?.response.message);
-    return {
-      props: {
-        data: null,
-      },
-    };
-  }
-}
 
 const CourseDash = async ({ params: { id } }: { params: { id: number } }) => {
   const { props } = await fetchCourseInfo(id);
