+++ app/(main)/c/[id]/t/[tid]/leaderboard/page.tsx
@@ -1,8 +1,5 @@
-import axios from 'axios';
-import { cookies } from 'next/headers';
 import Link from 'next/link';
-import { redirect } from 'next/navigation';
-import React from 'react';
+import { fetchLeaderboard } from './action';
 
 //todo: extract common leaderboard logic
 
@@ -110,41 +107,7 @@ const StudentCell = ({ name, link }: { name: string; link: string }) => {
   );
 };
 
-export const fetchLeaderboard = async (cid: number, tid: number) => {
-  try {
-    if (!cookies().get('access_token')) {
-      redirect('/auth/signin');
-      return {
-        data: null,
-      };
-    }
 
-    const response = await axios.get(
-      `http://localhost:5000/test/${tid}/leaderboard`,
-      {
-        headers: {
-          Authorization: `Bearer ${cookies().get('access_token')?.value}`,
-        },
-      }
-    );
-
-    const data = response.data;
-
-    return {
-      data,
-    };
-  } catch (err: any) {
-    console.log('error: ', err);
-    if (err.response.status === 404) redirect('/not-found');
-    else if (err.response.status === 401) redirect(`/c/${cid}`);
-    else if (err.response.status === 500) redirect(`/c/${cid}`);
-
-    return {
-      data: null,
-      status: err.response.status,
-    };
-  }
-};
 
 const LeaderboardPage = async ({
   params,
