+++ app/(test)/t/[id]/q/[qid]/action.ts
@@ -1,0 +1,29 @@
+'use server'
+
+import axios from "axios";
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+
+export const fetchQuestion = async (qid: number) => {
+  if (!cookies().get('access_token')) {
+    redirect('/auth/signin');
+    return {
+      data: null,
+    };
+  }
+
+  try {
+    const response = await axios.get(`http://localhost:5000/question/${qid}`, {
+      headers: {
+        Authorization: `Bearer ${cookies().get('access_token')?.value}`,
+      },
+    });
+
+    const data = response.data;
+    return { data };
+  } catch (err) {
+    console.log(err);
+    // todo: error handling
+    return { data: null };
+  }
+};
