+++ app/(test)/t/[id]/q/[qid]/page.tsx
@@ -1,9 +1,7 @@
 import { QuestionSideBar } from '@/components/Popups';
 import { EditorSection } from '@/components/Question/Solve/Editor/editorSection';
 import { InformationSection } from '@/components/Question/Solve/Information/informationSection';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
-import axios from 'axios';
+import { fetchQuestion } from '@/app/action';
 
 interface Question {
   id: number;
@@ -17,30 +15,6 @@ interface Question {
     explaination: string;
   }[];
 }
-
-export const fetchQuestion = async (qid: number) => {
-  if (!cookies().get('access_token')) {
-    redirect('/auth/signin');
-    return {
-      data: null,
-    };
-  }
-
-  try {
-    const response = await axios.get(`http://localhost:5000/question/${qid}`, {
-      headers: {
-        Authorization: `Bearer ${cookies().get('access_token')?.value}`,
-      },
-    });
-
-    const data = response.data;
-    return { data };
-  } catch (err) {
-    console.log(err);
-    // todo: error handling
-    return { data: null };
-  }
-};
 
 const QuestionPage = async ({
   params,
