--- components/Editor/index.ts
@@ -1,6 +1,0 @@
-export * from './editor';
-export * from './landing';
-export * from './outputWindow';
-export * from './outputDetails';
-export * from './languageDropDown';
-export * from './customInput';
